feat: selected arch
