+++ Selected/run.sh
@@ -1,0 +1,1 @@
+poetry run python Selected/src/main.py $@
