+++ Selected/src/models/list_tasks.py
@@ -1,0 +1,10 @@
+from typing import List as ListType
+from dataclasses import dataclass
+from Selected.src.models.lists import List
+from Selected.src.models.tasks import Task
+
+
+@dataclass
+class ListTasks:
+    list: List
+    tasks: ListType[Task]
