+++ Selected/src/views/base.py
