+++ Selected/src/views/tasks/select_parent_task_prompt.py
@@ -1,0 +1,19 @@
+from Selected.src.prompter import Prompter
+from Selected.src.views.base import View
+from typing import List as ListType, Optional
+from Selected.src.models.tasks import Task
+
+
+class SelectParentTaskPromptView(View):
+    def __init__(self, tasks: ListType[Task]):
+        self.tasks = tasks
+        self.prompter = Prompter(
+            "What is the parent task's title? (optional): ", False, str
+        )
+
+    def render(self) -> Optional[str]:
+        print("Available parent tasks:")
+        for task in self.tasks:
+            print("\t", task.title)
+        parent_task_name = self.prompter.ask()
+        return parent_task_name
