+++ Selected/src/views/tasks/toggle_prompt.py
@@ -1,0 +1,15 @@
+from Selected.src.prompter import Prompter
+from Selected.src.views.base import View
+
+
+class ToggleTaskPromptView(View):
+    def __init__(self):
+        self.title_prompter = Prompter(
+            "Enter a task title to change the completion status: ",
+            required=True,
+            input_type=str,
+        )
+
+    def render(self) -> str:
+        task_title = self.title_prompter.ask()
+        return task_title
