+++ Unselected/run.sh
@@ -1,0 +1,1 @@
+poetry run python Unselected/src/main.py
